quell a bunch of warnings, mostly int->float implicit cast warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18526 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDuiTypeIn.cxx
@@ -41,7 +41,7 @@ void		HUDuiTypeIn::setObfuscation(bool on)
   obfuscate = on;
 }
 
-int			HUDuiTypeIn::getMaxLength() const
+size_t			HUDuiTypeIn::getMaxLength() const
 {
   return maxLength;
 }
